server: create Handler object in new handlers pkg
Handler is responsible for wiring up HTTP handlers. This used to be
handled directly by the Server object, but this object will have other
responsiblities soon, so it makes sense to separate things out
