+++ server/main.go
@@ -22,12 +22,8 @@ func main() {
 		port = defaultPort
 	}
 
-	s, err := server.NewServer(port)
-	if err != nil {
-		log.Fatalf("Failed to create new Server object: %v\n", err)
-	}
-	s.Start([]server.RouteHandler{
-		server.NewHealthHandler(s),
-		server.NewWSHandler(s),
-	})
+	// Spin up the server and all of its goroutines.
+	s := server.NewServer()
+	err := s.Start(port)
+	log.Fatalf("failed to spin up the server: %v\n", err)
 }
